Add RadiusLocationModal component
