+++ front-end/pages/NewMessagePage.jsx/NewMessagePage.css
@@ -1,0 +1,3 @@
+.new-message-page {
+    padding: 10vh 5vh;
+}
