+++ front-end/pages/SingleConvoPage/SingleConvoPage.css
@@ -1,0 +1,3 @@
+.single-convo-page {
+    padding: 10vh 5vh;
+}
